requirements-dev.txt
PyYaml>=6.0.2,<7.0.0
pydantic>=2.11.4,<3.0.0
black>=25.1.0,<26.0.0
isort>=6.0.1,<7.0.0
ruff>=0.11.8,<1.0.0
pytest>=8.3.5,<9.0.0
pytest-cov>=6.1.1,<7.0.0
